requirements.txt
numpy
dlib
gpiozero
opencv-python
Pillow
pydantic
requests
pyserial